reorg api
