+++ siad/apipeer.go
@@ -1,0 +1,31 @@
+package main
+
+import (
+	"net/http"
+
+	"github.com/NebulousLabs/Sia/network"
+)
+
+func (d *daemon) peerAddHandler(w http.ResponseWriter, req *http.Request) {
+	err := d.core.AddPeer(network.Address(req.FormValue("addr")))
+	if err != nil {
+		http.Error(w, err.Error(), 400)
+		return
+	}
+
+	writeSuccess(w)
+}
+
+func (d *daemon) peerRemoveHandler(w http.ResponseWriter, req *http.Request) {
+	err := d.core.RemovePeer(network.Address(req.FormValue("addr")))
+	if err != nil {
+		http.Error(w, err.Error(), 400)
+		return
+	}
+
+	writeSuccess(w)
+}
+
+func (d *daemon) peerStatusHandler(w http.ResponseWriter, req *http.Request) {
+	writeJSON(w, d.core.AddressBook())
+}
